move PngImage from atlas-core to atlas-chart
The image utilities are only used for charting and this
makes it easier to get consistency with upcoming font
changes.
