+++ atlas-chart/src/main/scala/com/netflix/atlas/chart/util/PngImage.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.atlas.core.util
+package com.netflix.atlas.chart.util
 
 import java.awt.Color
 import java.awt.Font
@@ -28,10 +28,10 @@ import java.io.ByteArrayOutputStream
 import java.io.InputStream
 import java.io.OutputStream
 import java.text.AttributedString
-import javax.imageio.IIOImage
-import javax.imageio.ImageIO
 
 import com.sun.imageio.plugins.png.PNGMetadata
+import javax.imageio.IIOImage
+import javax.imageio.ImageIO
 
 object PngImage {
