+++ atlas-chart/src/test/resources/pngimage/test.png
