+++ atlas-chart/src/test/resources/pngimage/test_diff.png
