+++ atlas-chart/src/test/resources/pngimage/test_error_800x100.png
