+++ atlas-chart/src/test/resources/pngimage/test_system_error.png
