+++ atlas-chart/src/test/resources/pngimage/test_user_error.png
